refactor: rename EntityRootBase to AggregateBase
